requirements.txt
tqdm==4.62.3
torch==1.8.1
scipy==1.7.0
torchvision==0.9.1
numpy==1.19.5
qpth==0.0.15
pandas==1.1.5
Pillow==9.2.0
scikit_learn==1.1.1
